Поменял логику флага для видео

+++ overschool/courses/api_views/homework.py
@@ -137,17 +137,17 @@ class HomeworkViewSet(
         serializer.is_valid(raise_exception=True)
 
         video = request.FILES.get("video")
-        if video and video_use:
+        if video:
             if instance.video:
                 s3.delete_file(str(instance.video))
             base_lesson = BaseLesson.objects.get(homeworks=instance)
             serializer.validated_data["video"] = s3.upload_large_file(
                 request.FILES["video"], base_lesson
             )
-        elif not video:
+        elif not video and video_use:
             if instance.video:
                 s3.delete_file(str(instance.video))
-        else:
+        elif not video and not video_use:
             serializer.validated_data["video"] = instance.video
 
         self.perform_update(serializer)
